Remove comments, rename Navbar to Sidebar
